+++ src/main/java/de/samply/notification/OperationType.java
@@ -1,5 +1,8 @@
 package de.samply.notification;
 
 public enum OperationType {
-    GENERATE_TOKEN
+    SEND_QUERY_TO_BRIDGEHEAD_AND_EXECUTE,
+    SEND_QUERY_TO_BRIDGEHEAD,
+    CREATE_DATASHIELD_TOKEN
+
 }