+++ src/main/resources/db/migration/V001__initialize_schema_and_tables.sql
@@ -80,18 +80,6 @@ CREATE TABLE samply.project_document
     label             TEXT
 );
 
-CREATE TABLE samply.bridgehead_operation
-(
-    id          SERIAL PRIMARY KEY,
-    bridgehead  TEXT      NOT NULL,
-    user_email  TEXT      NOT NULL,
-    timestamp   TIMESTAMP NOT NULL,
-    http_status TEXT      NOT NULL,
-    error       TEXT,
-    type        TEXT      NOT NULL,
-    project_id  BIGINT    NOT NULL
-);
-
 CREATE TABLE samply.notification
 (
     id             SERIAL                   NOT NULL PRIMARY KEY,
@@ -101,7 +89,8 @@ CREATE TABLE samply.notification
     bridgehead     TEXT,
     operation_type TEXT                     NOT NULL,
     details        TEXT                     NOT NULL,
-    error          TEXT
+    error          TEXT,
+    http_status    TEXT
 );
 
 CREATE TABLE samply.notification_user_action
@@ -133,11 +122,6 @@ ALTER TABLE samply.project_document
         FOREIGN KEY (project_id)
             REFERENCES samply.project (id);
 
-ALTER TABLE samply.bridgehead_operation
-    ADD CONSTRAINT fk_project_id
-        FOREIGN KEY (project_id)
-            REFERENCES samply.project (id);
-
 ALTER TABLE samply.notification
     ADD CONSTRAINT fk_project_id
         FOREIGN KEY (project_id) REFERENCES samply.project (id);
@@ -150,6 +134,5 @@ CREATE INDEX idx_project_bridgehead_project_id ON samply.project_bridgehead (pro
 CREATE INDEX idx_project_bridgehead_user_project_bridgehead_id ON samply.project_bridgehead_user (project_bridgehead_id);
 CREATE INDEX idx_project_document_project_id ON samply.project_document (project_id);
 CREATE INDEX idx_project_query_id ON samply.project (query_id);
-CREATE INDEX idx_bridgehead_operation_project_id ON samply.bridgehead_operation (project_id);
 CREATE INDEX idx_notification_project_id ON samply.notification (project_id);
 CREATE INDEX idx_notification_id ON samply.notification_user_action (notification_id);