+++ CHANGELOG.md
@@ -79,3 +79,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch all user visible notifications
 - Accepted state requirement for setting user as developer or pilot
 - Notification User Action
+- Replace Bridgehead Operation through Notification